requirements.txt
absl-py==0.10.0
astunparse==1.6.3
attrs==21.2.0
cachetools==3.1.1
certifi==2021.5.30
charset-normalizer==2.0.4
cloudpickle==1.6.0
decorator==5.0.9
dill==0.3.4
dm-tree==0.1.6
flatbuffers==2.0
future==0.18.2
gast==0.4.0
google-auth==2.0.1
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
h5py==3.1.0
idna==3.2
importlib-resources==5.2.2
jax==0.2.19
jaxlib==0.1.70
keras-nightly==2.7.0.dev2021082507
Keras-Preprocessing==1.1.2
libclang==11.1.0
Markdown==3.3.4
mpmath==1.2.1
nest-asyncio==1.5.1
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
portpicker==1.3.9
promise==2.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scipy==1.7.1
semantic-version==2.8.5
six==1.15.0
tb-nightly==2.7.0a20210825
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-federated-nightly==0.19.0.dev20210825
tensorflow-io-gcs-filesystem==0.20.0
tensorflow-metadata==1.2.0
tensorflow-model-optimization==0.5.0
tensorflow-privacy==0.7.2
tensorflow-probability==0.13.0
termcolor==1.1.0
tf-estimator-nightly==2.7.0.dev2021082508
tf-nightly==2.7.0.dev20210825
tqdm==4.28.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
